requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.3.1
feedgen==1.0.0
pytz==2024.1
gunicorn==21.2.0